requirements.txt
accelerate==1.4.0
auraloss==0.4.0
einops==0.8.1
huggingface_hub==0.29.1
hydra-core==1.3.2
librosa==0.10.2.post1
matplotlib==3.10.7
nltk==3.9.1
nnAudio==0.3.3
numpy==2.2.6
omegaconf==2.3.0
pandas==2.3.3
phonemizer==3.3.0
Pillow==12.0.0
safetensors==0.6.2
scipy==1.15.3
soundfile==0.13.1
spaces==0.42.1
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
vocos==0.1.0